fix: 拆分alertItem，增加新的alert类型
