+++ packages/zent/src/alert/demos/styles.md
@@ -1,17 +1,19 @@
 ---
 order: 2
 zh-CN:
-  title: "四种样式：info, success, warning 和 error"
+  title: "五种样式：info, success, warning, error 和 slight"
 	info: 这个是默认的 info 样式
 	success: 这个是 success 样式
 	warning: 这个是 warning 样式
 	error: 这个是 error 样式
+	slight: 这个是 slight 样式
 en-US:
-  title: "Four styles: info, success, warning and error"
+  title: "Five styles: info, success, warning, error and slight"
 	info: Default info style
 	success: Success style
 	warning: Warning style
 	error: Error style
+	slight: Slight style
 ---
 
 ```jsx
@@ -23,6 +25,7 @@ ReactDOM.render(
 		<Alert type="success">{i18n.success}</Alert>
 		<Alert type="warning">{i18n.warning}</Alert>
 		<Alert type="error">{i18n.error}</Alert>
+		<Alert type="slight">{i18n.slight}</Alert>
 	</div>,
 	mountNode
 );
