+++ packages/zent/src/alert/index.ts
@@ -1,4 +1,5 @@
 import Alert from './Alert';
+export * from './AlertItem';
 export * from './Alert';
 export * from './types';
 export default Alert;
